Move personal details controller to a subfolder
